third_party/WebKit/LayoutTests/platform/win/http/tests/fetch/workers/thorough/redirect-password-other-https-expected.txt

CONSOLE WARNING: line 14: Subresource requests whose URLs contain embedded credentials (e.g. `https://user:pass@host/`) are deprecated, and will be blocked in M59, around June 2017. See https://www.chromestatus.com/feature/5669008342777856 for more details.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&method=GET. Redirect from 'http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&method=GET' has been blocked by CORS policy: Redirect location 'https://user@localhost:8443/serviceworker/resources/fetch-access-control.php?&ACAOrigin=*' contains a username and password, which is disallowed for cross-origin requests.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%3Apass%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&method=GET. Redirect from 'http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%3Apass%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&method=GET' has been blocked by CORS policy: Redirect location 'https://user:pass@localhost:8443/serviceworker/resources/fetch-access-control.php?&ACAOrigin=*' contains a username and password, which is disallowed for cross-origin requests.
CONSOLE ERROR: Fetch API cannot load https://localhost:8443/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Fuser%40127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*. Redirect from 'https://localhost:8443/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Fuser%40127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*' has been blocked by CORS policy: Redirect location 'http://user@127.0.0.1:8000/serviceworker/resources/fetch-access-control.php?ACAOrigin=*' contains a username and password, which is disallowed for cross-origin requests.
CONSOLE ERROR: Fetch API cannot load https://localhost:8443/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Fuser%3Apass%40127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*. Redirect from 'https://localhost:8443/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Fuser%3Apass%40127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*' has been blocked by CORS policy: Redirect location 'http://user:pass@127.0.0.1:8000/serviceworker/resources/fetch-access-control.php?ACAOrigin=*' contains a username and password, which is disallowed for cross-origin requests.
CONSOLE ERROR: Fetch API cannot load https://localhost:8443/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*. Redirect from 'https://localhost:8443/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*' has been blocked by CORS policy: Redirect location 'https://user@localhost:8443/serviceworker/resources/fetch-access-control.php?ACAOrigin=*' contains a username and password, which is disallowed for cross-origin requests.
CONSOLE ERROR: Fetch API cannot load https://localhost:8443/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%3Apass%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*. Redirect from 'https://localhost:8443/serviceworker/resources/redirect.php?Redirect=https%3A%2F%2Fuser%3Apass%40localhost%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&method=GET&ACAOrigin=*' has been blocked by CORS policy: Redirect location 'https://user:pass@localhost:8443/serviceworker/resources/fetch-access-control.php?ACAOrigin=*' contains a username and password, which is disallowed for cross-origin requests.
This is a testharness.js-based test.
PASS Startup 
PASS executeTest-0 
PASS executeTest-1 
PASS executeTest-2 
PASS executeTest-3 
PASS executeTest-4 
PASS executeTest-5 
PASS executeTest-6 
PASS executeTest-7 
PASS executeTest-8 
PASS executeTest-9 
PASS executeTest-10 
PASS executeTest-11 
PASS executeTest-12 
PASS executeTest-13 
PASS executeTest-14 
PASS executeTest-15 
PASS executeTest-16 
PASS executeTest-17 
Harness: the test ran to completion.
